fix 'missing IDs' bug by adding getter method

+++ src/main/java/com/example/mitarbeiterdb/repo/PersonRepo.java
@@ -1,12 +1,12 @@
 package com.example.mitarbeiterdb.repo;
 
+import com.example.mitarbeiterdb.entity.PersonEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.stereotype.Repository;
 
-import com.example.mitarbeiterdb.entity.PersonEntity;
 @EnableJpaRepositories
 @Repository
-public interface PersonRepo extends JpaRepository<PersonEntity, Integer> {
+public interface PersonRepo extends JpaRepository<PersonEntity, Long> {
 
 }
