+++ src/main/java/com/example/mitarbeiterdb/repo/StandortRepo.java
@@ -1,13 +1,12 @@
 package com.example.mitarbeiterdb.repo;
 
+import com.example.mitarbeiterdb.entity.StandortEntity;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.mitarbeiterdb.entity.StandortEntity;
 import org.springframework.stereotype.Repository;
 
 @Repository
 @EnableAutoConfiguration
-public interface StandortRepo extends JpaRepository<StandortEntity, Integer> {
+public interface StandortRepo extends JpaRepository<StandortEntity, Long> {
 
 }
